# Optimize tag query by avoiding multi-query