## Changes

Use SQL built-in aggregate string to retrieve journal row and tags in 1 go

### Latest `dev` Benchmarks 

|          Method |  TagMode |         Mean |      Error |     StdDev |       Gen0 |      Gen1 |    Allocated |
|---------------- |--------- |-------------:|-----------:|-----------:|-----------:|----------:|-------------:|
|    QueryByTag10 |      Csv | 1,789.544 ms | 13.2316 ms | 12.3768 ms |          - |         - |    449.88 KB |
|   QueryByTag100 |      Csv | 1,756.674 ms | 11.3694 ms | 10.6350 ms |          - |         - |    1368.1 KB |
|  QueryByTag1000 |      Csv | 1,753.439 ms |  9.9202 ms |  9.2793 ms |  1000.0000 |         - |  10675.91 KB |
| QueryByTag10000 |      Csv | 1,935.542 ms | 31.4495 ms | 29.4179 ms | 12000.0000 | 3000.0000 | 103178.35 KB |
|    QueryByTag10 | TagTable |     3.071 ms |  0.0377 ms |  0.0353 ms |    42.9688 |    3.9063 |    356.29 KB |
|   QueryByTag100 | TagTable |     4.798 ms |  0.0838 ms |  0.0784 ms |   156.2500 |   31.2500 |   1307.99 KB |
|  QueryByTag1000 | TagTable |    24.617 ms |  0.4469 ms |  0.4180 ms |  1312.5000 |  656.2500 |  10922.98 KB |
| QueryByTag10000 | TagTable |   420.234 ms |  7.7032 ms |  7.2056 ms | 12000.0000 | 3000.0000 | 106987.34 KB |

### This PR's Benchmarks

|          Method |  TagMode |         Mean |      Error |     StdDev |       Gen0 |      Gen1 |      Gen2 |    Allocated |
|---------------- |--------- |-------------:|-----------:|-----------:|-----------:|----------:|----------:|-------------:|
|    QueryByTag10 |      Csv | 1,760.393 ms | 27.1970 ms | 25.4401 ms |          - |         - |         - |    449.35 KB |
|   QueryByTag100 |      Csv | 1,766.355 ms | 25.0182 ms | 23.4021 ms |          - |         - |         - |   1368.33 KB |
|  QueryByTag1000 |      Csv | 1,755.960 ms | 33.8171 ms | 34.7276 ms |  1000.0000 |         - |         - |  10733.67 KB |
| QueryByTag10000 |      Csv | 1,905.026 ms | 22.3564 ms | 20.9122 ms | 12000.0000 | 3000.0000 |         - |  103175.1 KB |
|    QueryByTag10 | TagTable |     2.336 ms |  0.0389 ms |  0.0344 ms |    39.0625 |    3.9063 |         - |    341.51 KB |
|   QueryByTag100 | TagTable |     3.943 ms |  0.0705 ms |  0.0660 ms |   148.4375 |   31.2500 |         - |   1262.26 KB |
|  QueryByTag1000 | TagTable |    18.597 ms |  0.3570 ms |  0.3506 ms |  1281.2500 |  718.7500 |         - |  10481.17 KB |
| QueryByTag10000 | TagTable |   184.446 ms |  3.3447 ms |  2.9650 ms | 14000.0000 | 3666.6667 | 1666.6667 | 103154.93 KB |