@Arkatufus 2023-04-17
Another optimization I tried is by passing the original journal query into the tag query like so:
```
var tagRows = await connection
    .GetTable<JournalTagRow>()
    .Where(r => r.OrderingId.In(rowQuery.Select(jr => jr.Ordering)))
    .Select(r => new TagRow
    {
        OrderingId = r.OrderingId,
        TagValue = r.TagValue
    })
    .ToListAsync();
```

This actually degrades the performance:
| Tag Count |  TagMode |        Mean |     Error |    StdDev |       Gen0 |      Gen1 |      Gen2 |    Allocated |
|---------------- |--------- |------------:|----------:|----------:|-----------:|----------:|----------:|-------------:|
|    10 |      Csv | 1,752.93 ms | 18.677 ms | 15.596 ms |          - |         - |         - |    450.28 KB |
|   100 |      Csv | 1,722.95 ms | 11.452 ms |  8.941 ms |          - |         - |         - |   1368.69 KB |
|  1000 |      Csv | 1,726.12 ms |  9.396 ms |  8.789 ms |  1000.0000 |         - |         - |  10734.09 KB |
| 10000 |      Csv | 1,895.89 ms | 20.333 ms | 19.020 ms | 12000.0000 | 3000.0000 |         - |    103175 KB |
|    10 | TagTable |    20.60 ms |  0.208 ms |  0.194 ms |    31.2500 |         - |         - |    379.39 KB |
|   100 | TagTable |    23.35 ms |  0.466 ms |  0.518 ms |   156.2500 |   31.2500 |         - |   1320.66 KB |
|  1000 | TagTable |    43.39 ms |  0.850 ms |  0.795 ms |  1250.0000 |  666.6667 |         - |  10836.95 KB |
| 10000 | TagTable |   451.57 ms |  7.244 ms |  6.049 ms | 13000.0000 | 4000.0000 | 1000.0000 | 105942.11 KB |