@Arkatufus 2023-04-17
> @Arkatufus looks like this broke something across several Db implementations

Should be fixed now